Wav2Lip/requirements.txt
absl-py>=2.1.0
attrs>=24.2.0
audioread>=3.0.1
batch-face>=1.5.0
certifi>=2024.8.30
cffi>=1.17.1
charset-normalizer>=3.4.0
colorama>=0.4.6
contourpy>=1.3.1
cycler>=0.12.1
decorator>=5.1.1
ffmpeg-python>=0.2.0
filelock>=3.16.1
flatbuffers>=24.3.25
fonttools>=4.55.2
fsspec>=2024.10.0
future>=1.0.0
idna>=3.10
intel-openmp>=2021.4.0
jax>=0.4.36
jaxlib>=0.4.36
Jinja2>=3.1.4
joblib>=1.4.2
kiwisolver>=1.4.7
lazy_loader>=0.4
librosa>=0.10.2.post1
llvmlite>=0.43.0
MarkupSafe>=3.0.2
matplotlib>=3.9.3
mediapipe>=0.10.18
mkl>=2021.4.0
ml_dtypes>=0.5.0
mpmath>=1.3.0
msgpack>=1.1.0
networkx>=3.4.2
numba>=0.60.0
numpy>=1.26.4
opencv-contrib-python>=4.10.0.84
opencv-python>=4.10.0.84
opencv-transforms>=0.0.6
opt_einsum>=3.4.0
packaging>=24.2
pandas>=2.2.3
pillow>=11.0.0
pip>=24.2
platformdirs>=4.3.6
pooch>=1.8.2
protobuf>=4.25.5
pycparser>=2.22
pyparsing>=3.2.0
python-dateutil>=2.9.0.post0
pytz>=2024.2
requests>=2.32.3
scikit-learn>=1.5.2
scipy>=1.14.1
sentencepiece>=0.2.0
setuptools>=75.1.0
six>=1.17.0
sixdrepnet>=0.1.6
sounddevice>=0.5.1
soundfile>=0.12.1
soxr>=0.5.0.post1
sympy>=1.13.3
tbb>=2021.13.1
threadpoolctl>=3.5.0
torch>=2.3.0
torchvision>=0.18.0
tqdm>=4.67.1
typing_extensions>=4.12.2
tzdata>=2024.2
urllib3>=2.2.3
wheel>=0.44.0
